requirements.txt
pyrogram
tgcrypto
quart
google-generativeai
aiohttp